Update Program.cs
It is divided to tasks :
Task 1 => get name, id, age and department for a student from the user then display their data.
Task 2 => get name, ssn, age and salary for an employee from the user and check if age > 0 and salary > 20000 then display their data.
Task 3 => get a number from the user then check if the number is positive or negative or zero.
Task 4 => get a number from the user then check if the number is even or odd.
Task 5 => get rectangle length and width then check if it is a square or not.